+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Breaking changes
 
-## [Unreleased]
+## [1.0.0-beta.4]
 
 ### Added
 - feat: Add `--no-modifyTime` option, allowing to modify date but keeping original time information.